# backend/swaggerui/dist/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "To do app",
    "description": "todo application repo https://github.com/codescalersinternships/kassem-ToDo",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "https://virtserver.swaggerhub.com/OMARKASSEM099/ToDo-Go/1.0.0",
      "description": "SwaggerHub API Auto Mocking"
    }
  ],
  "paths": {
...
    "/todo{taskId}": {
      "get": {
        "summary": "get task by Id",
        "operationId": "getTask",
        "parameters": [
          {
            "name": "taskId",
            "in": "path",
            "required": true,
            "style": "simple",
            "explode": false,
            "schema": {
              "type": "integer",
              "example": 23
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/task"
                }
              }
            }
          },
          "400": {
            "description": "bad request"
          },
          "500": {
            "description": "faild due server error"
          }
        }
      },
      "put": {
        "summary": "update task by Id",
        "operationId": "updateTask",
        "parameters": [
          {
            "name": "taskId",
            "in": "path",
            "required": true,
            "style": "simple",
            "explode": false,
            "schema": {
              "type": "integer",
              "example": 23
            }
          }
        ],
        "responses": {
          "200": {
            "description": "updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/task"
                }
              }
            }
          },
          "400": {
            "description": "bad request"
          },
          "500": {
            "description": "faild due server error"
          }
        }
      },
      "delete": {
        "summary": "get task by Id",
        "operationId": "deleteTask",
        "parameters": [
          {
            "name": "taskId",
            "in": "path",
            "required": true,
            "style": "simple",
            "explode": false,
            "schema": {
              "type": "integer",
              "example": 23
            }
          }
        ],
        "responses": {
          "200": {
            "description": "deleted"
          },
          "400": {
            "description": "bad request"
          },
          "500": {
            "description": "faild due server error"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "task": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "example": 3
          },
          "task": {
            "type": "string",
            "example": "create tests"
          },
          "status": {
            "type": "boolean",
            "example": false
          }
        }
      },
...
    }
  }
}